Set Dependency  Injection service in the program class
