--- TranscriptManagement/RepositoryInterfaces/IWriteContext.cs
@@ -1,7 +1,0 @@
-﻿namespace TranscriptManagement.RepositoryInterfaces
-{
-    public interface IWriteContext
-    {
-        string SaveResult(string path, string matric);
-    }
-}
