+++ TranscriptManagerApp/TranscriptManagerClient.csproj
@@ -6,6 +6,11 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.9" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.9" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\TranscriptManagement\TranscriptManagement.csproj" />
     <ProjectReference Include="..\TranscriptManagerPersistence\TranscriptManagerPersistence.csproj" />
   </ItemGroup>
