+++ TranscriptManagerPersistence/Paths/FilePaths.cs
@@ -1,0 +1,8 @@
+﻿namespace TranscriptManagerPersistence.Paths
+{
+    public class FilePaths : IFilePaths
+    {
+        public string pathLecturer { get ; set; } = "LecturerList.txt";
+        public string pathStudent { get; set; } = @"C:\Users\HP\source\repos\Transcript\TranscriptManagerPersistence\Results";
+    }
+}
