+++ TranscriptManagerPersistence/Paths/IFilePaths.cs
@@ -1,0 +1,8 @@
+﻿namespace TranscriptManagerPersistence.Paths
+{
+    public interface IFilePaths
+    {
+        string pathLecturer { get; set; }
+        string pathStudent { get; set; }
+    }
+}
