--- TranscriptManagerPersistence/Repositories/IWriteContext.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace TranscriptManagerPersistence.Repositories
-{
-    class IWriteContext
-    {
-    }
-}
